+++ controller/editorcontroller.php
@@ -21,15 +21,14 @@ namespace OCA\Onlyoffice\Controller;
 
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\ContentSecurityPolicy;
+use OCP\AppFramework\Http\DataDownloadResponse;
 use OCP\AppFramework\Http\RedirectResponse;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\Constants;
 use OCP\Files\File;
-use OCP\Files\Folder;
 use OCP\Files\ForbiddenException;
 use OCP\Files\IRootFolder;
 use OCP\Files\NotPermittedException;
-use OCP\Files\Storage\IPersistentLockingStorage;
 use OCP\IL10N;
 use OCP\ILogger;
 use OCP\IRequest;
@@ -38,6 +37,7 @@ use OCP\IURLGenerator;
 use OCP\IUser;
 use OCP\IUserSession;
 use OCP\Share\IManager;
+use OCP\Share;
 
 use OCA\Files\Helper;
 
@@ -118,6 +118,13 @@ class EditorController extends Controller {
     private $versionManager;
 
     /**
+     * Share manager
+     *
+     * @var IManager
+     */
+    private $shareManager;
+
+    /**
      * Mobile regex from https://github.com/ONLYOFFICE/CommunityServer/blob/v9.1.1/web/studio/ASC.Web.Studio/web.appsettings.config#L35
      */
     const USER_AGENT_MOBILE = "/android|avantgo|playbook|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od|ad)|iris|kindle|lge |maemo|midp|mmp|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\\/|plucker|pocket|psp|symbian|treo|up\\.(browser|link)|vodafone|wap|windows (ce|phone)|xda|xiino/i";
@@ -156,6 +163,7 @@ class EditorController extends Controller {
         $this->logger = $logger;
         $this->config = $config;
         $this->crypt = $crypt;
+        $this->shareManager = $shareManager;
 
         $this->versionManager = new VersionManager($AppName, $root);
 
@@ -214,7 +222,7 @@ class EditorController extends Controller {
             $this->logger->error("Folder for file creation was not found: $dir", ["app" => $this->appName]);
             return ["error" => $this->trans->t("The required folder was not found")];
         }
-        if (!$folder->isCreatable()) {
+        if (!($folder->isCreatable() && $folder->isUpdateable())) {
             $this->logger->error("Folder for file creation without permission: $dir", ["app" => $this->appName]);
             return ["error" => $this->trans->t("You don't have enough permission to create")];
         }
@@ -222,7 +230,10 @@ class EditorController extends Controller {
         if (empty($templateId)) {
             $template = TemplateManager::GetEmptyTemplate($name);
         } else {
-            $template = TemplateManager::GetTemplate($templateId);
+            $templateFile = TemplateManager::GetTemplate($templateId);
+            if ($templateFile) {
+                $template = $templateFile->getContent();
+            }
         }
 
         if (!$template) {
@@ -255,22 +266,148 @@ class EditorController extends Controller {
      *
      * @param string $name - file name
      * @param string $dir - folder path
+     * @param string $templateId - file identifier
      * 
      * @return TemplateResponse|RedirectResponse
      *
      * @NoAdminRequired
      * @NoCSRFRequired
      */
-    public function createNew($name, $dir) {
+    public function createNew($name, $dir, $templateId = null) {
         $this->logger->debug("Create from editor: $name in $dir", ["app" => $this->appName]);
 
-        $result = $this->create($name, $dir);
+        $result = $this->create($name, $dir, $templateId);
         if (isset($result["error"])) {
             return $this->renderError($result["error"]);
         }
 
         $openEditor = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".editor.index", ["fileId" => $result["id"]]);
         return new RedirectResponse($openEditor);
+    }
+
+    /**
+     * Get users
+     *
+     * @return array
+     *
+     * @NoAdminRequired
+     * @NoCSRFRequired
+     */
+    public function users() {
+        $this->logger->debug("Search users", ["app" => $this->appName]);
+        $result = [];
+
+        if (!$this->config->isUserAllowedToUse()) {
+            return $result;
+        }
+
+        $userId = $this->userSession->getUser()->getUID();
+        $users = \OC::$server->getUserManager()->search("");
+        foreach ($users as $user) {
+            $email = $user->getEMailAddress();
+            if ($user->getUID() != $userId
+                && !empty($email)) {
+                array_push($result, [
+                    "email" => $email,
+                    "name" => $user->getDisplayName()
+                ]);
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Send notify about mention
+     *
+     * @param int $fileId - file identifier
+     * @param string $anchor - the anchor on target content
+     * @param string $comment - comment
+     * @param array $emails - emails array to whom to send notify
+     *
+     * @return array
+     *
+     * @NoAdminRequired
+     * @NoCSRFRequired
+     */
+    public function mention($fileId, $anchor, $comment, $emails) {
+        $this->logger->debug("mention: from $fileId to " . json_encode($emails), ["app" => $this->appName]);
+
+        if (!$this->config->isUserAllowedToUse()) {
+            return ["error" => $this->trans->t("Not permitted")];
+        }
+
+        if (empty($emails)) {
+            return ["error" => $this->trans->t("Failed to send notification")];
+        }
+
+        $recipientIds = [];
+        foreach ($emails as $email) {
+            $recipients = \OC::$server->getUserManager()->getByEmail($email);
+            foreach ($recipients as $recipient) {
+                $recipientId = $recipient->getUID(); 
+                if (!in_array($recipientId, $recipientIds)) {
+                    array_push($recipientIds, $recipientId);
+                }
+            }
+        }
+
+        $user = $this->userSession->getUser();
+        $userId = null;
+        if (!empty($user)) {
+            $userId = $user->getUID();
+        }
+
+        list ($file, $error, $share) = $this->getFile($userId, $fileId);
+        if (isset($error)) {
+            $this->logger->error("Mention: $fileId $error", ["app" => $this->appName]);
+            return ["error" => $this->trans->t("Failed to send notification")];
+        }
+
+        $notificationManager = \OC::$server->getNotificationManager();
+        $notification = $notificationManager->createNotification();
+        $notification->setApp($this->appName)
+            ->setDateTime(new \DateTime())
+            ->setObject("mention", $comment)
+            ->setSubject("mention_info", [
+                "notifierId" => $userId,
+                "fileId" => $file->getId(),
+                "fileName" => $file->getName(),
+                "anchor" => $anchor
+            ]);
+
+        $canShare = ($file->getPermissions() & Constants::PERMISSION_SHARE) === Constants::PERMISSION_SHARE;
+
+        $accessList = [];
+        foreach ($this->shareManager->getSharesByPath($file) as $share) {
+            array_push($accessList, $share->getSharedWith());
+        }
+
+        foreach ($recipientIds as $recipientId) {
+            if (!in_array($recipientId, $accessList)) {
+                if (!$canShare) {
+                    continue;
+                }
+
+                $share = $this->shareManager->newShare();
+                $share->setNode($file)
+                    ->setShareType(Share::SHARE_TYPE_USER)
+                    ->setSharedBy($userId)
+                    ->setSharedWith($recipientId)
+                    ->setShareOwner($userId)
+                    ->setPermissions(Constants::PERMISSION_READ);
+
+                $this->shareManager->createShare($share);
+
+                $this->logger->debug("mention: share $fileId to $recipientId", ["app" => $this->appName]);
+            }
+
+            $notification->setUser($recipientId);
+
+            $notificationManager->notify($notification);
+        }
+
+        return ["message" => $this->trans->t("Notification sent successfully")];
     }
 
     /**
@@ -344,7 +481,7 @@ class EditorController extends Controller {
         }
 
         $folder = $file->getParent();
-        if (!$folder->isCreatable()) {
+        if (!($folder->isCreatable() && $folder->isUpdateable())) {
             $folder = $this->root->getUserFolder($userId);
         }
 
@@ -404,7 +541,7 @@ class EditorController extends Controller {
             $this->logger->error("Folder for saving file was not found: $dir", ["app" => $this->appName]);
             return ["error" => $this->trans->t("The required folder was not found")];
         }
-        if (!$folder->isCreatable()) {
+        if (!($folder->isCreatable() && $folder->isUpdateable())) {
             $this->logger->error("Folder for saving file without permission: $dir", ["app" => $this->appName]);
             return ["error" => $this->trans->t("You don't have enough permission to create")];
         }
@@ -452,6 +589,10 @@ class EditorController extends Controller {
      */
     public function history($fileId, $shareToken = null) {
         $this->logger->debug("Request history for: $fileId", ["app" => $this->appName]);
+
+        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
+            return ["error" => $this->trans->t("Not permitted")];
+        }
 
         $history = [];
 
@@ -569,6 +710,10 @@ class EditorController extends Controller {
      */
     public function version($fileId, $version, $shareToken = null) {
         $this->logger->debug("Request version for: $fileId ($version)", ["app" => $this->appName]);
+
+        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
+            return ["error" => $this->trans->t("Not permitted")];
+        }
 
         $version = empty($version) ? null : $version;
 
@@ -700,6 +845,98 @@ class EditorController extends Controller {
     }
 
     /**
+     * Download method
+     *
+     * @param int $fileId - file identifier
+     * @param string $toExtension - file extension to download
+     * @param bool $template - file is template
+     *
+     * @return DataDownloadResponse|TemplateResponse
+     *
+     * @NoAdminRequired
+     * @NoCSRFRequired
+     */
+    public function download($fileId, $toExtension = null, $template = false) {
+        $this->logger->debug("Download: $fileId $toExtension", ["app" => $this->appName]);
+
+        if (!$this->config->isUserAllowedToUse()) {
+            return $this->renderError($this->trans->t("Not permitted"));
+        }
+
+        if ($template) {
+            $templateFile = TemplateManager::GetTemplate($fileId);
+
+            if (empty($templateFile)) {
+                $this->logger->info("Download: template not found: $fileId", ["app" => $this->appName]);
+                return $this->renderError($this->trans->t("File not found"));
+            }
+
+            $file = $templateFile;
+        } else {
+            $user = $this->userSession->getUser();
+            $userId = null;
+            if (!empty($user)) {
+                $userId = $user->getUID();
+            }
+
+            list ($file, $error, $share) = $this->getFile($userId, $fileId);
+
+            if (isset($error)) {
+                $this->logger->error("Download: $fileId $error", ["app" => $this->appName]);
+                return $this->renderError($error);
+            }
+        }
+
+        $fileStorage = $file->getStorage();
+        if ($fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage")) {
+            $storageShare = $fileStorage->getShare();
+            if (method_exists($storageShare, "getAttributes")) {
+                $attributes = $storageShare->getAttributes();
+
+                $permissionsDownload = $attributes->getAttribute("permissions", "download");
+                if ($permissionsDownload !== null && $permissionsDownload !== true) {
+                    return $this->renderError($this->trans->t("Not permitted"));
+                }
+            }
+        }
+
+        $fileName = $file->getName();
+        $ext = strtolower(pathinfo($fileName, PATHINFO_EXTENSION));
+        $toExtension = strtolower($toExtension);
+
+        if ($toExtension === null
+            || $ext === $toExtension
+            || $template) {
+            return new DataDownloadResponse($file->getContent(), $fileName, $file->getMimeType());
+        }
+
+        $newFileUri = null;
+        $documentService = new DocumentService($this->trans, $this->config);
+        $key = $this->fileUtility->getKey($file);
+        $fileUrl = $this->getUrl($file, $user);
+        try {
+            $newFileUri = $documentService->GetConvertedUri($fileUrl, $ext, $toExtension, $key);
+        } catch (\Exception $e) {
+            $this->logger->logException($e, ["message" => "GetConvertedUri: " . $file->getId(), "app" => $this->appName]);
+            return $this->renderError($e->getMessage());
+        }
+
+        try {
+            $newData = $documentService->Request($newFileUri);
+        } catch (\Exception $e) {
+            $this->logger->logException($e, ["message" => "Failed to download converted file", "app" => $this->appName]);
+            return $this->renderError($this->trans->t("Failed to download converted file"));
+        }
+
+        $fileNameWithoutExt = substr($fileName, 0, strlen($fileName) - strlen($ext) - 1);
+        $newFileName = $fileNameWithoutExt . "." . $toExtension;
+
+        $formats = $this->config->FormatsSetting();
+
+        return new DataDownloadResponse($newData, $newFileName, $formats[$toExtension]["mime"]);
+    }
+
+    /**
      * Print editor section
      *
      * @param integer $fileId - file identifier
@@ -708,13 +945,14 @@ class EditorController extends Controller {
      * @param integer $version - file version
      * @param bool $inframe - open in frame
      * @param bool $template - file is template
+     * @param string $anchor - anchor for file content
      *
      * @return TemplateResponse|RedirectResponse
      *
      * @NoAdminRequired
      * @NoCSRFRequired
      */
-    public function index($fileId, $filePath = null, $shareToken = null, $version = 0, $inframe = false, $template = false) {
+    public function index($fileId, $filePath = null, $shareToken = null, $version = 0, $inframe = false, $template = false, $anchor = null) {
         $this->logger->debug("Open: $fileId ($version) $filePath", ["app" => $this->appName]);
 
         if (empty($shareToken) && !$this->userSession->isLoggedIn()) {
@@ -742,7 +980,8 @@ class EditorController extends Controller {
             "shareToken" => $shareToken,
             "version" => $version,
             "template" => $template,
-            "inframe" => false
+            "inframe" => false,
+            "anchor" => $anchor
         ];
 
         if ($inframe === true) {
@@ -782,287 +1021,6 @@ class EditorController extends Controller {
      */
     public function PublicPage($fileId, $shareToken, $version = 0, $inframe = false) {
         return $this->index($fileId, null, $shareToken, $version, $inframe);
-    }
-
-    /**
-     * Get template loader Onlyoffice
-     *
-     * @return TemplateResponse
-     *
-     * @NoAdminRequired
-     * @NoCSRFRequired
-     * @PublicPage
-     */
-    public function loader() {
-        return new TemplateResponse($this->appName, "loader", [], "plain");
-    }
-
-    /**
-     * Collecting the file parameters for the document service
-     *
-     * @param integer $fileId - file identifier
-     * @param string $filePath - file path
-     * @param string $shareToken - access token
-     * @param integer $version - file version
-     * @param bool $inframe - open in frame
-     * @param bool $desktop - desktop label
-     * @param bool $template - file is template
-     *
-     * @return array
-     *
-     * @NoAdminRequired
-     * @PublicPage
-     */
-    public function config($fileId, $filePath = null, $shareToken = null, $version = 0, $inframe = false, $desktop = false, $template = false) {
-
-        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
-            return ["error" => $this->trans->t("Not permitted")];
-        }
-
-        $user = $this->userSession->getUser();
-        $userId = null;
-        if (!empty($user)) {
-            $userId = $user->getUID();
-        }
-
-        list ($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId, $filePath, $template) : $this->fileUtility->getFileByToken($fileId, $shareToken);
-
-        if (isset($error)) {
-            $this->logger->error("Config: $fileId $error", ["app" => $this->appName]);
-            return ["error" => $error];
-        }
-
-        $fileName = $file->getName();
-        $ext = strtolower(pathinfo($fileName, PATHINFO_EXTENSION));
-        $format = $this->config->FormatsSetting()[$ext];
-        if (!isset($format)) {
-            $this->logger->info("Format is not supported for editing: $fileName", ["app" => $this->appName]);
-            return ["error" => $this->trans->t("Format is not supported")];
-        }
-
-        $fileUrl = $this->getUrl($file, $user, $shareToken, $version, null, $template);
-
-        $key = null;
-        if ($version > 0
-            && $this->versionManager->available) {
-            $owner = $file->getFileInfo()->getOwner();
-            if ($owner !== null) {
-                $versions = array_reverse($this->versionManager->getVersionsForFile($owner, $file->getFileInfo()));
-
-                if ($version <= count($versions)) {
-                    $fileVersion = array_values($versions)[$version - 1];
-
-                    $key = $this->fileUtility->getVersionKey($fileVersion);
-                }
-            }
-        }
-        if ($key === null) {
-            $key = $this->fileUtility->getKey($file, true);
-        }
-        $key = DocumentService::GenerateRevisionId($key);
-
-        $params = [
-            "document" => [
-                "fileType" => $ext,
-                "key" => $key,
-                "permissions" => [],
-                "title" => $fileName,
-                "url" => $fileUrl,
-            ],
-            "documentType" => $format["type"],
-            "editorConfig" => [
-                "lang" => str_replace("_", "-", \OC::$server->getL10NFactory("")->get("")->getLanguageCode())
-            ]
-        ];
-
-        $restrictedEditing = false;
-        $fileStorage = $file->getStorage();
-        if (empty($shareToken) && $fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage")) {
-            $storageShare = $fileStorage->getShare();
-            if (method_exists($storageShare, "getAttributes"))
-            {
-                $attributes = $storageShare->getAttributes();
-
-                $permissionsDownload = $attributes->getAttribute("permissions", "download");
-                if ($permissionsDownload !== null) {
-                    $params["document"]["permissions"]["download"] = $params["document"]["permissions"]["print"] = $params["document"]["permissions"]["copy"] = $permissionsDownload === true;
-                }
-
-                if (isset($format["review"]) && $format["review"]) {
-                    $permissionsReviewOnly = $attributes->getAttribute($this->appName, "review");
-                    if ($permissionsReviewOnly !== null && $permissionsReviewOnly === true) {
-                        $restrictedEditing = true;
-                        $params["document"]["permissions"]["review"] = true;
-                    }
-                }
-
-                if (isset($format["fillForms"]) && $format["fillForms"]) {
-                    $permissionsFillFormsOnly = $attributes->getAttribute($this->appName, "fillForms");
-                    if ($permissionsFillFormsOnly !== null && $permissionsFillFormsOnly === true) {
-                        $restrictedEditing = true;
-                        $params["document"]["permissions"]["fillForms"] = true;
-                    }
-                }
-
-                if (isset($format["comment"]) && $format["comment"]) {
-                    $permissionsCommentOnly = $attributes->getAttribute($this->appName, "comment");
-                    if ($permissionsCommentOnly !== null && $permissionsCommentOnly === true) {
-                        $restrictedEditing = true;
-                        $params["document"]["permissions"]["comment"] = true;
-                    }
-                }
-
-                if (isset($format["modifyFilter"]) && $format["modifyFilter"]) {
-                    $permissionsModifyFilter = $attributes->getAttribute($this->appName, "modifyFilter");
-                    if ($permissionsModifyFilter !== null) {
-                        $params["document"]["permissions"]["modifyFilter"] = $permissionsModifyFilter === true;
-                    }
-                }
-            }
-        }
-
-        $isPersistentLock = false;
-        if ($version < 1
-            && (\OC::$server->getConfig()->getAppValue("files", "enable_lock_file_action", "no") === "yes")
-            && $fileStorage->instanceOfStorage(IPersistentLockingStorage::class)) {
-
-            $locks = $fileStorage->getLocks($file->getFileInfo()->getInternalPath(), false);
-            if (count($locks) > 0) {
-                $activeLock = $locks[0];
-                $lockOwner = explode(' ', trim($activeLock->getOwner()))[0];
-                if ($userId !== $lockOwner) {
-                    $isPersistentLock = true;
-                    $this->logger->debug("File $fileId is locked by $lockOwner", ["app" => $this->appName]);
-                }
-            }
-        }
-
-        $canEdit = isset($format["edit"]) && $format["edit"];
-        $editable = $version < 1
-                    && !$template
-                    && $file->isUpdateable()
-                    && !$isPersistentLock
-                    && (empty($shareToken) || ($share->getPermissions() & Constants::PERMISSION_UPDATE) === Constants::PERMISSION_UPDATE);
-        $params["document"]["permissions"]["edit"] = $editable;
-        if (($editable || $restrictedEditing) && $canEdit) {
-            $ownerId = null;
-            $owner = $file->getOwner();
-            if (!empty($owner)) {
-                $ownerId = $owner->getUID();
-            }
-
-            $hashCallback = $this->crypt->GetHash(["userId" => $userId, "ownerId" => $ownerId, "fileId" => $file->getId(), "filePath" => $filePath, "shareToken" => $shareToken, "action" => "track"]);
-            $callback = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".callback.track", ["doc" => $hashCallback]);
-
-            if (!empty($this->config->GetStorageUrl())) {
-                $callback = str_replace($this->urlGenerator->getAbsoluteURL("/"), $this->config->GetStorageUrl(), $callback);
-            }
-
-            $params["editorConfig"]["callbackUrl"] = $callback;
-        } else {
-            $params["editorConfig"]["mode"] = "view";
-        }
-
-        if (\OC::$server->getRequest()->isUserAgent([$this::USER_AGENT_MOBILE])) {
-            $params["type"] = "mobile";
-        }
-
-        if (!empty($userId)) {
-            $params["editorConfig"]["user"] = [
-                "id" => $this->buildUserId($userId),
-                "name" => $user->getDisplayName()
-            ];
-        }
-
-        $folderLink = null;
-
-        if (!empty($shareToken)) {
-            $node = $share->getNode();
-            if ($node instanceof Folder) {
-                $sharedFolder = $node;
-                $folderPath = $sharedFolder->getRelativePath($file->getParent()->getPath());
-                if (!empty($folderPath)) {
-                    $linkAttr = [
-                        "path" => $folderPath,
-                        "scrollto" => $file->getName(),
-                        "token" => $shareToken
-                    ];
-                    $folderLink = $this->urlGenerator->linkToRouteAbsolute("files_sharing.sharecontroller.showShare", $linkAttr);
-                }
-            }
-        } else if (!empty($userId)) {
-            $userFolder = $this->root->getUserFolder($userId);
-            $folderPath = $userFolder->getRelativePath($file->getParent()->getPath());
-            if (!empty($folderPath)) {
-                $linkAttr = [
-                    "dir" => $folderPath,
-                    "scrollto" => $file->getName()
-                ];
-                $folderLink = $this->urlGenerator->linkToRouteAbsolute("files.view.index", $linkAttr);
-            }
-
-            switch($params["documentType"]) {
-                case "text":
-                    $createName = $this->trans->t("Document") . ".docx";
-                    break;
-                case "spreadsheet":
-                    $createName = $this->trans->t("Spreadsheet") . ".xlsx";
-                    break;
-                case "presentation":
-                    $createName = $this->trans->t("Presentation") . ".pptx";
-                    break;
-            }
-
-            $createParam = [
-                "dir" => "/",
-                "name" => $createName
-            ];
-
-            if (!empty($folderPath)) {
-                $folder = $userFolder->get($folderPath);
-                if (!empty($folder) && $folder->isCreatable()) {
-                    $createParam["dir"] = $folderPath;
-                }
-            }
-
-            $createUrl = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".editor.create_new", $createParam);
-
-            $params["editorConfig"]["createUrl"] = urldecode($createUrl);
-        }
-
-        if ($folderLink !== null) {
-            $params["editorConfig"]["customization"]["goback"] = [
-                "url"  => $folderLink
-            ];
-
-            if (!$desktop) {
-                if ($this->config->GetSameTab()) {
-                    $params["editorConfig"]["customization"]["goback"]["blank"] = false;
-                    if ($inframe === true) {
-                        $params["editorConfig"]["customization"]["goback"]["requestClose"] = true;
-                    }
-                }
-            }
-        }
-
-        if ($inframe === true) {
-            $params["_files_sharing"] = \OC::$server->getAppManager()->isEnabledForUser("files_sharing");
-        }
-
-        $params = $this->setCustomization($params);
-
-        if ($this->config->UseDemo()) {
-            $params["editorConfig"]["tenant"] = $this->config->GetSystemValue("instanceid", true);
-        }
-
-        if (!empty($this->config->GetDocumentServerSecret())) {
-            $token = \Firebase\JWT\JWT::encode($params, $this->config->GetDocumentServerSecret());
-            $params["token"] = $token;
-        }
-
-        $this->logger->debug("Config is generated for: $fileId ($version) with key $key", ["app" => $this->appName]);
-
-        return $params;
     }
 
     /**
@@ -1172,86 +1130,6 @@ class EditorController extends Controller {
         $instanceId = $this->config->GetSystemValue("instanceid", true);
         $userId = $instanceId . "_" . $userId;
         return $userId;
-    }
-
-    /**
-     * Set customization parameters
-     *
-     * @param array params - file parameters
-     *
-     * @return array
-     */
-    private function setCustomization($params) {
-        //default is true
-        if ($this->config->GetCustomizationChat() === false) {
-            $params["editorConfig"]["customization"]["chat"] = false;
-        }
-
-        //default is false
-        if ($this->config->GetCustomizationCompactHeader() === true) {
-            $params["editorConfig"]["customization"]["compactHeader"] = true;
-        }
-
-        //default is false
-        if ($this->config->GetCustomizationFeedback() === true) {
-            $params["editorConfig"]["customization"]["feedback"] = true;
-        }
-
-        //default is false
-        if ($this->config->GetCustomizationForcesave() === true) {
-            $params["editorConfig"]["customization"]["forcesave"] = true;
-        }
-
-        //default is true
-        if ($this->config->GetCustomizationHelp() === false) {
-            $params["editorConfig"]["customization"]["help"] = false;
-        }
-
-        //default is original
-        $reviewDisplay = $this->config->GetCustomizationReviewDisplay();
-        if ($reviewDisplay !== "original") {
-            $params["editorConfig"]["customization"]["reviewDisplay"] = $reviewDisplay;
-        }
-
-        //default is false
-        if ($this->config->GetCustomizationToolbarNoTabs() === true) {
-            $params["editorConfig"]["customization"]["toolbarNoTabs"] = true;
-        }
-
-
-        /* from system config */
-
-        $autosave = $this->config->GetSystemValue($this->config->_customization_autosave);
-        if (isset($autosave)) {
-            $params["editorConfig"]["customization"]["autosave"] = $autosave;
-        }
-
-        $customer = $this->config->GetSystemValue($this->config->_customization_customer);
-        if (isset($customer)) {
-            $params["editorConfig"]["customization"]["customer"] = $customer;
-        }
-
-        $loaderLogo = $this->config->GetSystemValue($this->config->_customization_loaderLogo);
-        if (isset($loaderLogo)) {
-            $params["editorConfig"]["customization"]["loaderLogo"] = $loaderLogo;
-        }
-
-        $loaderName = $this->config->GetSystemValue($this->config->_customization_loaderName);
-        if (isset($loaderName)) {
-            $params["editorConfig"]["customization"]["loaderName"] = $loaderName;
-        }
-
-        $logo = $this->config->GetSystemValue($this->config->_customization_logo);
-        if (isset($logo)) {
-            $params["editorConfig"]["customization"]["logo"] = $logo;
-        }
-
-        $zoom = $this->config->GetSystemValue($this->config->_customization_zoom);
-        if (isset($zoom)) {
-            $params["editorConfig"]["customization"]["zoom"] = $zoom;
-        }
-
-        return $params;
     }
 
     /**