+++ controller/settingscontroller.php
@@ -152,7 +152,7 @@ class SettingsController extends Controller {
                                     ) {
         $error = null;
         if (!$this->config->SelectDemo($demo === true)) {
-            $error = $this->trans->t("The 30-day test period is over, you can no longer connect to demo ONLYOFFICE Document Server.");
+            $error = $this->trans->t("The 30-day test period is over, you can no longer connect to demo ONLYOFFICE Docs server.");
         }
         if ($demo !== true) {
             $this->config->SetDocumentServerUrl($documentserver);
@@ -275,6 +275,18 @@ class SettingsController extends Controller {
      * @return array
      */
     private function GetGlobalTemplates() {
-        return TemplateManager::GetGlobalTemplates();
+        $templates = [];
+        $templatesList = TemplateManager::GetGlobalTemplates();
+
+        foreach ($templatesList as $templateItem) {
+            $template = [
+                "id" => $templateItem->getId(),
+                "name" => $templateItem->getName(),
+                "type" => TemplateManager::GetTypeTemplate($templateItem->getMimeType())
+            ];
+            array_push($templates, $template);
+        }
+
+        return $templates;
     }
 }